Separated files for settings & render & converted class compinent to functional
